+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.0.0
 
+- Added Bun as a supported package manager.
 - Updated action to Node v20.
 - Updated dependencies.
   - `@actions/github` v5 -> v6